fix: Remove duplicate code, fix nav scrolling
