+++ ietf/static/js/ietf.js
@@ -24,8 +24,6 @@ if (!process.env.BUILD_DEPLOY) {
 
 import Cookies from "js-cookie";
 
-import debounce from "lodash/debounce";
-
 import { populate_nav } from "./nav.js";
 
 // setup CSRF protection using jQuery
@@ -204,18 +202,6 @@ $(function () {
                 extraNav.children().appendTo('#righthand-extra');
                 extraNav.remove();
             }
-
-            $(document)
-                // Chrome apparently wants this debounced to something >10ms,
-                // otherwise the main view doesn't scroll?
-                .on("scroll", debounce(function () {
-                    const item = $('#righthand-nav')
-                        .find(".active")
-                        .last();
-                    if (item.length) {
-                        item[0].scrollIntoView({ block: "center", behavior: "smooth" });
-                    }
-                }, 100));
 
             // offset the scrollspy to account for the menu bar
             const contentOffset = contentElement ? contentElement.offset().top : 0;
